Mostly housecleaning:
* Moving SFML to resources folder and integrating it into sln (to eliminate the bazillion compilation warnings about missing pdbs)
* Adding boost to resources folder so that we can use it without worrying about having it locally installed
* Removing display.h/display.cpp (including IPC code) from Algovis C++ project
* Eliminating other warnings (mainly dll export warnings)
* Getting rid of VS2010 sln since it's outdated and neither of us are using '10
* Creating a docs folder for each project which needs it
* Doing TODOs


git-svn-id: https://svn.ece.auckland.ac.nz/svn/p4p-2010-g79@6 a75856b6-468a-0410-9532-ba8c368bb4fd

+++ branches/trunk/view/src/display.cpp
@@ -69,7 +69,7 @@ private:
 			float ySize = 0;
 			float x = address.GetRect().Right + xGap;
 			std::vector<std::string>& container = i->second;
-			for (int j = 0; j < container.size(); j++)
+			for (unsigned j = 0; j < container.size(); j++)
 			{
 				sf::String text(container[j].c_str(), font);
 				text.SetColor(sf::Color(255, 255, 255));
@@ -111,7 +111,7 @@ private:
 				{
 					int width = Event.Size.Width;
 					int height = Event.Size.Height;
-					win->SetView(sf::View(sf::FloatRect(0, 0, width, height)));
+					win->SetView(sf::View(sf::FloatRect(0, 0, (float) width, (float) height)));
 					glViewport(0, 0, width, height);
 				}
